+++ src/GraphDataContext.tsx
@@ -30,6 +30,13 @@ import {
 } from "./GraphManager/hooks/useCreateUser";
 import { LoginFn, useLogin } from "./GraphManager/hooks/useLoginUser";
 
+//// TODO(skep): remove "translated"-types, the backend always receives the
+// language header, thus translation should happen there, to avoid large amount
+// of unnecessary data to be transferred.
+//import { Node, Link } from "./GraphManager/GraphRenderer";
+//export type TranslatedNode = Node;
+//export type TranslatedGraphData = Link;
+
 export interface TranslatedNode {
   id: string;
   description: Text;
